requirements.txt
Flask
google-cloud-storage
google-cloud-aiplatform
vertexai
gunicorn
